Moving Question class to questions model package
